requirements.txt
aiofiles==23.1.0
aiogrpc==1.8
aniso8601==9.0.1
anyio==3.6.2
apturl==0.5.2
argcomplete==3.0.8
async-tkinter-loop==0.7.0
asyncio==3.4.3
attrs==23.1.0
autopep8==2.0.2
bcrypt==3.2.0
bleach==6.0.0
blinker==1.4
boto3==1.26.131
botocore==1.29.131
Brlapi==0.8.3
CacheControl==0.12.11
Cerberus==1.3.4
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
click-log==0.4.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.7
coverage==7.2.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
deprecation==2.1.0
distro==1.7.0
distro-info===1.1build1
docutils==0.19
dotty-dict==1.3.1
duplicity==0.8.21
empy==3.3.4
fasteners==0.14.1
Flask==2.2.3
Flask-RESTful==0.3.9
fonttools==4.39.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
gotrue==1.0.1
grpcio==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.20.2
httpx==0.23.3
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.12.0
invoke==1.7.3
iso3166==2.1.1
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kconfiglib==14.1.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
mapbox==0.18.1
Markdown==3.3.6
MarkupSafe==2.1.2
matplotlib==3.7.1
mavsdk==1.4.5
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.5
netifaces==0.11.0
numpy==1.24.3
nunavut==2.0.9
oauthlib==3.2.0
olefile==0.46
packaging==23.1
pandas==2.0.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pkgconfig==1.5.5
pkginfo==1.9.6
polyline==2.0.0
postgrest==0.10.6
priority==2.0.0
protobuf==3.12.4
psutil==5.9.5
ptyprocess==0.7.0
pybind11==2.9.1
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
pydantic==1.10.7
pydsdl==1.18.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymavlink==2.4.37
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyros-genmsg==0.5.8
pyrsistent==0.19.3
pyserial==3.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-gitlab==3.14.0
python-semantic-release==7.33.2
pytz==2022.1
pyulog==1.0.1
pyxdg==0.27
PyYAML==5.4.1
quart==0.18.4
readme-renderer==37.3
realtime==1.0.0
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
s3transfer==0.6.1
SecretStorage==3.3.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
storage3==0.5.2
StrEnum==0.4.10
supabase==1.0.3
supafunc==0.2.2
sympy==1.11.1
systemd-python==234
tk==0.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.65.0
ttystatus==0.38
twine==3.8.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
websockets==10.4
Werkzeug==2.2.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
